+++ src/main/kotlin/lean4ij/actions/ToggleLeanInfoViewInternal.kt
@@ -8,7 +8,12 @@ import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.CommonDataKeys
 import com.intellij.openapi.components.service
 import com.intellij.openapi.editor.Editor
+import com.intellij.openapi.editor.actions.AbstractToggleUseSoftWrapsAction
+import com.intellij.openapi.editor.impl.softwrap.SoftWrapAppliancePlaces
 import com.intellij.openapi.util.IconLoader.getIcon
+import kotlinx.coroutines.TimeoutCancellationException
+import kotlinx.coroutines.runBlocking
+import kotlinx.coroutines.withTimeout
 import lean4ij.infoview.LeanInfoViewWindowFactory
 import lean4ij.infoview.LeanInfoviewService
 import lean4ij.infoview.external.JcefInfoviewService
@@ -179,4 +184,37 @@ class ToggleLeanInfoviewToolbarVisibility : AnAction() {
         component.isVisible = !component.isVisible
     }
 
+}
+
+/**
+ * TODO all the actions should be nicely grouped
+ *      ref: https://github.com/asciidoctor/asciidoctor-intellij-plugin/pull/222
+ */
+class ToggleInternalInfoviewSoftWrap : AbstractToggleUseSoftWrapsAction(SoftWrapAppliancePlaces.MAIN_EDITOR, false) {
+
+    init {
+        templatePresentation.icon = getIcon("/icons/newLine.svg", javaClass);
+
+        // TODO not sure if it should call the method copy from, but
+        //      it make the text into 'Soft-Wrap', which is wrong and cannot distinguish from the builtin action
+        // copyFrom(ActionManager.getInstance().getAction(IdeActions.ACTION_EDITOR_USE_SOFT_WRAPS));
+    }
+
+    override fun getActionUpdateThread(): ActionUpdateThread {
+        return ActionUpdateThread.BGT;
+    }
+
+    override fun getEditor(e: AnActionEvent): Editor? {
+        val project = e.project ?: return null
+        val toolWindow = project.service<LeanInfoviewService>().toolWindow ?: return null
+        return runBlocking {
+            try {
+                withTimeout(1000) {
+                    toolWindow.getEditor()
+                }
+            } catch (ex: TimeoutCancellationException) {
+                null
+            }
+        }
+    }
 }