+++ src/main/kotlin/lean4ij/sdk/Lean4StdLibraryProvider.kt
@@ -7,6 +7,7 @@ import java.nio.file.Files
 import java.nio.file.Path
 import kotlin.io.path.isDirectory
 import kotlin.io.path.name
+import kotlin.io.path.notExists
 
 /**
  * from julia-intellij, check src/org/ice1000/julia/lang/module/julia-sdks.kt
@@ -14,7 +15,11 @@ import kotlin.io.path.name
 class Lean4StdLibraryProvider : AdditionalLibraryRootsProvider() {
     override fun getAdditionalProjectLibraries(project: Project): Collection<LeanLibrary> {
         val basePath = project.basePath ?: return listOf()
-        return Files.list(Path.of(basePath, ".lake", "packages"))
+        val packagesPath = Path.of(basePath, ".lake", "packages")
+        if (packagesPath.notExists()) {
+            return listOf()
+        }
+        return Files.list(packagesPath)
             .filter { it.isDirectory() }
             .map {
                 LeanLibrary(it.name, VfsUtil.findFile(it, true)!!)